Remove UNIX_EPOCH, unused and could encourage wrong UNIX <-> TAI translation
Users of the module might mistakenly think this constant implies UNIX time is
a constant 10 seconds behind TAI time. This has not been the case since 1972,
when the first leap second occured.

Since UNIXEPOCH isn't necessary for the module to function it's safer to
remove it. If a solid use case is presented then it or something like it could
be re-introduced.